+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/Arrow.kt
@@ -1,17 +1,16 @@
-package io.github.pylonmc.pylon.core.item.interfaces
+package io.github.pylonmc.pylon.core.item.base
 
 import com.destroystokyo.paper.event.player.PlayerReadyArrowEvent
 import org.bukkit.event.entity.EntityShootBowEvent
 
-@FunctionalInterface
-interface Bow {
+interface Arrow {
     /**
      * Called when the arrow is shot from the bow of any entity
      */
-    fun onBowReady(event: PlayerReadyArrowEvent)
+    fun onArrowReady(event: PlayerReadyArrowEvent)
 
     /**
      * Called when the arrow is shot from the bow of any entity
      */
-    fun onBowFired(event: EntityShootBowEvent)
+    fun onArrowShotFromBow(event: EntityShootBowEvent)
 }
