+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/BlockInteractor.kt
@@ -1,8 +1,7 @@
-package io.github.pylonmc.pylon.core.item.interfaces
+package io.github.pylonmc.pylon.core.item.base
 
 import org.bukkit.event.player.PlayerInteractEvent
 
-@FunctionalInterface
 interface BlockInteractor {
     /**
      * Called when a player right clicks a block while holding the item
