+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/Bow.kt
@@ -1,17 +1,16 @@
-package io.github.pylonmc.pylon.core.item.interfaces
+package io.github.pylonmc.pylon.core.item.base
 
 import com.destroystokyo.paper.event.player.PlayerReadyArrowEvent
 import org.bukkit.event.entity.EntityShootBowEvent
 
-@FunctionalInterface
-interface Arrow {
+interface Bow {
     /**
      * Called when the arrow is shot from the bow of any entity
      */
-    fun onArrowReady(event: PlayerReadyArrowEvent)
+    fun onBowReady(event: PlayerReadyArrowEvent)
 
     /**
      * Called when the arrow is shot from the bow of any entity
      */
-    fun onArrowShotFromBow(event: EntityShootBowEvent)
+    fun onBowFired(event: EntityShootBowEvent)
 }
