+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/Consumable.kt
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.core.item.interfaces
+package io.github.pylonmc.pylon.core.item.base
 
 import org.bukkit.event.player.PlayerItemConsumeEvent
