+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/Damageable.kt
@@ -1,10 +1,9 @@
-package io.github.pylonmc.pylon.core.item.interfaces
+package io.github.pylonmc.pylon.core.item.base
 
 import org.bukkit.event.player.PlayerItemBreakEvent
 import org.bukkit.event.player.PlayerItemDamageEvent
 import org.bukkit.event.player.PlayerItemMendEvent
 
-@FunctionalInterface
 interface Damageable {
     /**
      * Called when the item is damaged
