+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/FurnaceFuel.kt
@@ -1,8 +1,7 @@
-package io.github.pylonmc.pylon.core.item.interfaces
+package io.github.pylonmc.pylon.core.item.base
 
 import org.bukkit.event.inventory.FurnaceBurnEvent
 
-@FunctionalInterface
 interface FurnaceFuel {
     /**
      * Called when the item is burnt as fuel in a furnace, smoker, etc
